Move launcher prerequisites validation to separate file
